Remove extra hooks
